codex(web): extract db schema core
